+++ packages/nest-auth/src/lib/core/providers/github-auth.provider.ts
@@ -5,11 +5,11 @@ import { GITHUB_AUTH_PROVIDER } from '../../auth.constants';
 import { NestAuthUser } from '../../user/entities/user.entity';
 import { NestAuthIdentity } from '../../user/entities/identity.entity';
 import { IAuthModuleOptions } from '../interfaces/auth-module-options.interface';
-import { BaseAuthProvider } from './base-auth.provider';
+import { SocialAuthProvider } from './social-auth.provider';
 import { SocialCredentialsDto } from '../../auth/dto/credentials/social-credentials.dto';
 
 @Injectable()
-export class GitHubAuthProvider extends BaseAuthProvider {
+export class GitHubAuthProvider extends SocialAuthProvider {
     providerName = GITHUB_AUTH_PROVIDER;
     private githubConfig: IAuthModuleOptions['github'];